requirements.txt
asgiref==3.3.4
CacheControl==0.12.6
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
crypto==1.4.1
cryptography==3.4.7
Django==3.2.3
firebase==3.0.1
firebase-admin==5.0.0
flake8==3.9.2
gcloud==0.18.3
google-api-core==1.26.3
google-api-python-client==2.4.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-firestore==2.1.1
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.37.1
gunicorn==20.1.0
heroku==0.1.4
httplib2==0.19.1
jwcrypto==0.8
mccabe==0.6.1
msgpack==1.0.2
Naked==0.1.31
oauth2client==4.1.3
packaging==20.9
proto-plus==1.18.1
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
pyparsing==2.4.7
python-dateutil==1.5
python-jwt==3.3.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.1
sseclient==0.0.27
uritemplate==3.0.1
urllib3==1.26.4